Added chatResponse and onlinePlayers validation using class-transformer-validator on the backend
