+++ packages/backend/package.json
@@ -57,6 +57,7 @@
     "@types/supertest": "^2.0.8",
     "bcryptjs": "^2.4.3",
     "class-transformer": "^0.2.3",
+    "class-transformer-validator": "^0.8.0",
     "class-validator": "^0.11.1",
     "eslint-config-airbnb-base": "^14.1.0",
     "express": "^4.17.1",
