--- packages/backend/src/auth/online-players/dto/online-players.dto.ts
@@ -1,13 +1,0 @@
-import { IsDefined, IsString } from 'class-validator';
-
-// empty
-export enum OnlinePlayerRewards {}
-
-export class OnlinePlayer {
-  @IsString()
-  @IsDefined()
-  username!: string;
-
-  @IsDefined()
-  rewards!: OnlinePlayerRewards[];
-}
